Conflicts:
	lib/rspec-system/node_set.rb
	lib/rspec-system/node_set/base.rb
	lib/rspec-system/node_set/vagrant.rb
	lib/rspec-system/node_set/vsphere.rb
	lib/rspec-system/spec_helper.rb

+++ lib/rspec-system/node_set/vagrant.rb
@@ -14,7 +14,8 @@ module RSpecSystem
     # @param setname [String] name of the set to instantiate
     # @param config [Hash] nodeset configuration hash
     # @param custom_prefabs_path [String] path of custom prefabs yaml file
-    def initialize(setname, config, custom_prefabs_path)
+    # @param options [Hash] options Hash
+    def initialize(setname, config, custom_prefabs_path, options)
       super
       @vagrant_path = File.expand_path(File.join(RSpec.configuration.system_tmp, 'vagrant_projects', setname))
     end
@@ -53,8 +54,12 @@ module RSpecSystem
         v.close unless v.closed?
       end
 
-      log.info "[Vagrant#teardown] Running 'vagrant destroy'"
-      vagrant("destroy --force")
+      if destroy
+        log.info "[Vagrant#teardown] Running 'vagrant destroy'"
+        vagrant("destroy --force")
+      else
+        log.info "[Vagrant#teardown] Skipping 'vagrant destroy'"
+      end
       nil
     end
